--- src/components/ui/calendar/calendar.tsx
@@ -1,6 +1,0 @@
-import {useState} from "react";
-
-export function CalendarDemo() {
-    const [date, setDate] = useState<Date | undefined>(new Date())
-
-}
