--- src/components/ui/calendar/index.ts
@@ -1,1 +1,0 @@
-export * from './calendar'
