--- src/components/ui/date-picker/index.ts
@@ -1,1 +1,0 @@
-export * from './date-picker'
