--- src/components/ui/popover/index.ts
@@ -1,1 +1,0 @@
-export * from './popover'
